# cmd/swag/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "API Documentation Ecommerce Backend SHOPDEVGO",
        "termsOfService": "github.com/anonystick/go-ecommerce-backend-go",
        "contact": {
            "name": "TEAM TIPSGO",
            "url": "github.com/anonystick/go-ecommerce-backend-go",
            "email": "tipsgo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8002",
    "basePath": "/v1",
    "paths": {
        "/auth/register": {
            "post": {
                "description": "When user is registered send otp to email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth management"
                ],
                "summary": "User Registration",
                "parameters": [
                    {
                        "description": "payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authcommandrequest.UserRegistratorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponseData"
                        }
                    }
                }
            }
        },
        "/auth/save-account": {
            "post": {
                "description": "When user has registered send otp to email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth management"
                ],
                "summary": "User Base Registration",
                "parameters": [
                    {
                        "description": "payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authcommandrequest.SaveAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponseData"
                        }
                    }
                }
            }
        },
        "/auth/verify-account": {
            "post": {
                "description": "When user is verified otp from email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth management"
                ],
                "summary": "Verify OTP",
                "parameters": [
                    {
                        "description": "payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authcommandrequest.VerifyOTPRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponseData"
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "description": "Returns user details data based on ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get user details by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseData"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponseData"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "authcommandrequest.SaveAccountRequest": {
            "type": "object",
            "required": [
                "confirm_pass",
                "email",
                "name",
                "password",
                "token"
            ],
            "properties": {
                "birthday": {
                    "description": "Ngày sinh (optional), định dạng yyyy-mm-dd",
                    "type": "string"
                },
                "confirm_pass": {
                    "description": "Xác nhận password phải giống Password",
                    "type": "string"
                },
                "email": {
                    "description": "Email bắt buộc và phải đúng định dạng",
                    "type": "string"
                },
                "gender": {
                    "description": "Giới tính: 0=male, 1=female, 2=other",
                    "type": "integer",
                    "enum": [
                        0,
                        1,
                        2
                    ]
                },
                "name": {
                    "description": "Tên người dùng",
                    "type": "string"
                },
                "password": {
                    "description": "Password tối thiểu 8 ký tự",
                    "type": "string",
                    "minLength": 8
                },
                "phone": {
                    "description": "Số điện thoại (optional, theo chuẩn E.164)",
                    "type": "string"
                },
                "token": {
                    "description": "token đăng kí thông tin người dùng",
                    "type": "string"
                }
            }
        },
        "authcommandrequest.UserRegistratorRequest": {
            "type": "object",
            "required": [
                "email",
                "purpose"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "purpose": {
                    "description": "TEST_USER, CUSTOMER, ADMIN, etc.",
                    "type": "string"
                }
            }
        },
        "authcommandrequest.VerifyOTPRequest": {
            "type": "object",
            "required": [
                "email",
                "otp"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "otp": {
                    "type": "string"
                }
            }
        },
        "response.ErrorResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "detail": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.ResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}